+++ Makefile
@@ -1,0 +1,4 @@
+.PHONY: libsds
+
+libsds:
+	nim c --app:lib --mm:refc --outdir:build library/libsds.nim
